feat: move generate logic to its own crate
