+++ cli/generate/src/build_tables/build_lex_table.rs
@@ -6,7 +6,7 @@ use std::{
 use log::info;
 
 use super::{coincident_tokens::CoincidentTokenIndex, token_conflicts::TokenConflictMap};
-use crate::generate::{
+use crate::{
     dedup::split_state_id_groups,
     grammars::{LexicalGrammar, SyntaxGrammar},
     nfa::{CharacterSet, NfaCursor},
