+++ cli/generate/src/dsl.js
