+++ cli/generate/src/grammar-schema.json
