+++ cli/generate/src/nfa.rs
