+++ cli/generate/src/prepare_grammar/unicode-categories.json
