+++ cli/generate/src/prepare_grammar/unicode-category-aliases.json
