+++ cli/generate/src/prepare_grammar/unicode-property-aliases.json
