+++ cli/generate/src/render.rs
