+++ cli/generate/src/rules.rs
