+++ cli/generate/src/tables.rs
