+++ cli/src/lib.rs
@@ -1,7 +1,6 @@
 #![doc = include_str!("../README.md")]
 
 pub mod fuzz;
-pub mod generate;
 pub mod highlight;
 pub mod init;
 pub mod logger;
