+++ cli/src/tests/corpus_test.rs
@@ -14,7 +14,6 @@ use crate::{
         EDIT_COUNT, EXAMPLE_EXCLUDE, EXAMPLE_INCLUDE, ITERATION_COUNT, LANGUAGE_FILTER,
         LOG_GRAPH_ENABLED, START_SEED,
     },
-    generate,
     parse::perform_edit,
     test::{parse_tests, print_diff, print_diff_key, strip_sexp_fields},
     tests::{
@@ -353,8 +352,8 @@ fn test_feature_corpus_files() {
             grammar_path = test_path.join("grammar.json");
         }
         let error_message_path = test_path.join("expected_error.txt");
-        let grammar_json = generate::load_grammar_file(&grammar_path, None).unwrap();
-        let generate_result = generate::generate_parser_for_grammar(&grammar_json);
+        let grammar_json = tree_sitter_generate::load_grammar_file(&grammar_path, None).unwrap();
+        let generate_result = tree_sitter_generate::generate_parser_for_grammar(&grammar_json);
 
         if error_message_path.exists() {
             if EXAMPLE_INCLUDE.is_some() || EXAMPLE_EXCLUDE.is_some() {
