+++ cli/src/tests/parser_test.rs
@@ -4,6 +4,7 @@ use std::{
 };
 
 use tree_sitter::{IncludedRangesError, InputEdit, LogType, Parser, Point, Range};
+use tree_sitter_generate::{generate_parser_for_grammar, load_grammar_file};
 use tree_sitter_proc_macro::retry;
 
 use super::helpers::{
@@ -13,7 +14,6 @@ use super::helpers::{
 };
 use crate::{
     fuzz::edits::Edit,
-    generate::{generate_parser_for_grammar, load_grammar_file},
     parse::perform_edit,
     tests::{helpers::fixtures::fixtures_dir, invert_edit},
 };
